requirements.txt
chris_plugin==0.1.2
pycivet==0.1.2
loguru==0.6.0
pybicpl==0.5.1
numpy~=1.22
